refactor construction selection router
